Implementing control of theta parameters.
Allowing user to parameterize the population mutation rates (i.e.,
thetas) for the ancestral and two descendant populations for each
taxon pair. This commit incorporates the needed changes to
`msprior.c` and `msprior.h` to allow the user to specify a string
of three numbers that controls how many free theta parameters are
estimated (from 1 ['000'] to 3 ['012']) for each taxon pair.