Implementing `bottleProportionShared` option.
Updating `msprior.c` and `msprior.h` to give user the ability
to control whether the descendant populations of each taxon pair
share the same bottleneck severity parameter or not.

This required incorporating the `bottleProportionShared` variable and
associated default value, and adding some simple if/then logic
when the bottleneck proportions are drawn.