Fixing bug in `rejectinC.c`.
The `msReject` binary sometimes did not return the expected number
of posterior samples (i.e., tolerance * number of prior samples).

This was due to a bug in `rejectinC.c` that caused the vector
of draws with the smallest Euclidean distances to not get updated
correctly, resulting in empty strings being stored for some samples.
Basically, the loop that shifts draws down this vector when a draw
with a smaller distance is found was beginning at the wrong index
(0 rather than 1), which was causing it to be skipped under certain
conditions.

This commit fixes this bug.