Overhauling low-level string structs in `eureject`.
Adding `assign_c_array` function and making `s_array` an array
of `c_array`s. This allows book-keeping of both the number of
strings and the length of each of them within an `s_array`.
This solves a memory leak that was happening in `append_s_array`,
because memory was being allocated for new strings, without the
memory for the previous string being freed. Now, an `s_array`
gets populated with `c_array`s during `init_s_array`, and the
`c_arrays` are only `realloc`ed if needed in `append_s_array`.

Changing code throughout to accommodate the changes to `s_array`.