Properly bundling hashtab.
As per my last commit, there were duplicated versions of the
hashtab code (`hashtab.c` and `hashtab.h`) in the msBayes
distribution. This commit makes all code dependent on the
same copies that are bundled in the `hashtab` subdirectory.

This obviates the need for the duplicates, which will be removed
in the next commit.