Adding more GSL support to configure scripts.
`configure.ac` now fully supports the `--with-gsl` argument. When
`with-gsl=yes` it will look for GSL in several different ways/places.
When `with-gsl=no` it will not build GSL-dependent tools (`dpdraw`).
When a location is specified, it will first look for `gsl-config`,
and then resort to manually setting up basic GSL configurations
based on the location (if it exists).