Updating summary reporting in `eureject`.
Fixing bug in summary reporting in `eureject` introduced when
adding sample sizes to summary file output. I am adding a separate
variable back in that keeps track of how many samples were used
to calculate means and std deviations. This way, the summary will
report 0 when the means and std deviations are provided.