Increasing size of line buffer for eureject.
Increasing the size of the c_array used as a line buffer by eureject for
reading files. Increasing from 2^14 to 2^20.

Also, Updating CMakeLists.txt cmake file for tests to look for the
subunit library and include directory and updating the
include_directories and C_LIBS accordingly. This was necessary due to an
update in GCC (I think).

Also updating some tests of array_utils to report some more info.

All tests passing.